Add files via upload
HOTEL BOOKING PROJECT is online hotel room booking website.
This project aims at creating on Hotel Booking Management System which can be used by Admin and Customers. The admin to advise/publish the availability of rooms in different fare and customers are checking the availability of room in required fare. Customers should be able to know the availability of the rooms on a particular date to reserve in hotel. They should be able to reserve the available rooms according to their need in advance to make their stay comfortable. The admin hands the booking information of customers. The users can register and log into the system.
HTML, CSS, JAVASCRIPT, JSP, MYSQL.